Fixed bugs

+++ VortexMethod/Utils.hpp
@@ -5,24 +5,24 @@
 
 using std::vector;
 
-double round_(const double& num, const int n_places)
+float round_(const float& num, const int n_places)
 {
     return round(num * pow(10, n_places)) / pow(10, n_places);
 }
 
-double arctan(double& angle)
+float arctan(float& angle)
 {
     return std::atan(angle);
 }
 
-double square(double& num)
+float square(float& num)
 {
     return num * num;
 }
 
-double trapz(const std::vector<double>& y, double dx)
+float trapz(const std::vector<float>& y, float dx)
 {
-    double integral = 0.0;
+    float integral = 0.0;
     for (size_t i = 1; i < y.size(); ++i)
     {
         integral += 0.5 * dx * (y[i] + y[i - 1]);
@@ -30,9 +30,9 @@ double trapz(const std::vector<double>& y, double dx)
     return integral;
 }
 
-double trapz(const vector<double>& y, const vector<double> x)
+float trapz(const vector<float>& y, const vector<float> x)
 {
-    double integral = 0.0f;
+    float integral = 0.0f;
     for (size_t i = 1; i < y.size(); ++i) {
         //std::cout << "x[i]: " << x[i] << ", x[i-1]: " << x[i - 1] << ", y[i]: " << y[i] << ", y[i-1]: " << y[i - 1] << std::endl;
         integral += (x[i] - x[i - 1]) * (y[i] + y[i - 1]);
@@ -40,9 +40,9 @@ double trapz(const vector<double>& y, const vector<double> x)
     return integral * 0.5;
 }
 
-double trapz(const vector<vector<double>>& y, int second_index, vector<double>& x)
+float trapz(const vector<vector<float>>& y, int second_index, vector<float>& x)
 {
-    double integral = 0.0f;
+    float integral = 0.0f;
     for (size_t i = 1; i < y.size(); ++i) {
         //std::cout << "x[i]: " << x[i] << ", x[i-1]: " << x[i - 1] << ", y[i]: " << y[i] << ", y[i-1]: " << y[i - 1] << std::endl;
         integral += (x[i] - x[i - 1]) * (y[i][second_index] + y[i - 1][second_index]);
@@ -50,44 +50,44 @@ double trapz(const vector<vector<double>>& y, int second_index, vector<double>& 
     return integral * 0.5;
 }
 
-double h1_ig(const double L_ig, const double LHS, double mu, double alpha)
+float h1_ig(const float L_ig, const float LHS, float mu, float alpha)
 {
     return LHS - 2 * L_ig * sqrt(pow(mu, 2) + pow(mu * tan(abs(alpha)) + L_ig, 2));
 }
 
-double h2_ig(const double& L_ig, const double& LHS, const double& mu, const double& alpha)
+float h2_ig(const float& L_ig, const float& LHS, const float& mu, const float& alpha)
 {
-    const double term2 = sqrt(pow(mu, 2) + pow(mu * tan(abs(alpha)) + L_ig, 2));
-    const double diff = - (2 * term2 + (2 * L_ig * (mu * tan(abs(alpha)) + L_ig) / term2));
+    const float term2 = sqrt(pow(mu, 2) + pow(mu * tan(abs(alpha)) + L_ig, 2));
+    const float diff = - (2 * term2 + (2 * L_ig * (mu * tan(abs(alpha)) + L_ig) / term2));
     return diff;
 }
 
-double h1_i(const double& L_i, const double& L_ig_calc, const double& mu, const double& angle, const double& radius, const double& R_tip)
+float h1_i(const float& L_i, const float& L_ig_calc, const float& mu, const float& angle, const float& radius, const float& R_tip)
 {
-    const double LHS = (L_i) / (L_ig_calc);
-    const double RHS = 1 + ((1.333 * mu) / (1.2 * (L_i + L_ig_calc) + mu)) * ((radius * cos(angle)) / R_tip);
+    const float LHS = (L_i) / (L_ig_calc);
+    const float RHS = 1 + ((1.333 * mu) / (1.2 * (L_i + L_ig_calc) + mu)) * ((radius * cos(angle)) / R_tip);
     return LHS - RHS;
 }
 
-double h2_i(const double& L_i, const double& L_ig_calc, const double& mu, const double& angle, const double& radius, const double& R_tip)
+float h2_i(const float& L_i, const float& L_ig_calc, const float& mu, const float& angle, const float& radius, const float& R_tip)
 {
     return (1 / L_ig_calc) + ((1.333 * 1.2 * mu / pow(1.2 * (L_i + L_ig_calc) + mu, 2)) * radius * cos(angle) / R_tip);
 }
 
-double sumVecElem(const vector<double>& vec)
+float sumVecElem(const vector<float>& vec)
 {
-    double sum = 0.0f;
-    for (double el : vec) sum += el;
+    float sum = 0.0f;
+    for (float el : vec) sum += el;
     return sum;
 }
 
-vector<double> elementWiseMul(const vector<double>& a, const vector<double>& b)
+vector<float> elementWiseMul(const vector<float>& a, const vector<float>& b)
 {
     if (a.size() != b.size())
     {
         return {};
     }
-    vector<double> prod;
+    vector<float> prod;
     for (int i = 0; i < a.size(); i++)
     {
         prod.push_back(a[i] * b[i]);
@@ -95,9 +95,9 @@ vector<double> elementWiseMul(const vector<double>& a, const vector<double>& b)
     return prod;
 }
 
-vector<double> linspace(double start, double end, int num) {
-    vector<double> result(num);
-    double step = (end - start) / (num - 1);
+vector<float> linspace(float start, float end, int num) {
+    vector<float> result(num);
+    float step = (end - start) / (num - 1);
 
     for (int i = 0; i < num; ++i) {
         result[i] = start + i * step;
@@ -106,13 +106,13 @@ vector<double> linspace(double start, double end, int num) {
     return result;
 }
 
-vector<double> vectorAdd(const vector<double>& a, const vector<double>& b)
+vector<float> vectorAdd(const vector<float>& a, const vector<float>& b)
 {
     if (a.size() != b.size())
     {
         return {};
     }
-    vector<double> sum;
+    vector<float> sum;
     for (int i = 0; i < a.size(); i++)
     {
         sum.push_back(a[i] - b[i]);
@@ -120,13 +120,13 @@ vector<double> vectorAdd(const vector<double>& a, const vector<double>& b)
     return sum;
 }
 
-vector<double> vectorSub(vector<double>& a, vector<double>& b)
+vector<float> vectorSub(vector<float>& a, vector<float>& b)
 {
     if (a.size() != b.size())
     {
         return {};
     }
-    vector<double> diff(a.size(), 0);
+    vector<float> diff(a.size(), 0);
     for (int i = 0; i < a.size(); i++)
     {
         diff[i] = a[i] - b[i];
@@ -134,10 +134,10 @@ vector<double> vectorSub(vector<double>& a, vector<double>& b)
     return diff;
 }
 
-vector<vector<double>> vectorSub(vector<vector<double>>& a, vector<vector<double>>& b)
+vector<vector<float>> vectorSub(vector<vector<float>>& a, vector<vector<float>>& b)
 {
     if (a.size() != b.size()) return {};
-    vector<vector<double>> diff;
+    vector<vector<float>> diff;
     for (int i = 0; i < a.size(); i++)
     {
         diff.push_back(vectorSub(a[i], b[i]));
@@ -145,11 +145,11 @@ vector<vector<double>> vectorSub(vector<vector<double>>& a, vector<vector<double
     return diff;
 }
 
-//vector<vector<double>> scalarMul(const double&k, vector<double)
+//vector<vector<float>> scalarMul(const float&k, vector<float)
 
-//vector<double> scalarMul(const double& k, const vector<double>& vec)
+//vector<float> scalarMul(const float& k, const vector<float>& vec)
 //{
-//    vector<double> res;
+//    vector<float> res;
 //    for (auto& elem : vec)
 //    {
 //        res.push_back(k * elem);
@@ -157,13 +157,13 @@ vector<vector<double>> vectorSub(vector<vector<double>>& a, vector<vector<double
 //    return res;
 //}
 
-vector<double> vectorDiv(vector<double>& dividend, vector<double>& divisor)
+vector<float> vectorDiv(vector<float>& dividend, vector<float>& divisor)
 {
     if (dividend.size() != divisor.size())
     {
         return {};
     }
-    vector<double> res(dividend.size(), 0);
+    vector<float> res(dividend.size(), 0);
     for (int i = 0; i < dividend.size(); i++)
     {
         res[i] = dividend[i] / divisor[i];
@@ -172,7 +172,7 @@ vector<double> vectorDiv(vector<double>& dividend, vector<double>& divisor)
 }
 
 template<typename T>
-vector<T> scalarMul(const double& k, const vector<T>& vec)
+vector<T> scalarMul(const float& k, const vector<T>& vec)
 {
     std::vector<T> res;
     for (const auto& elem : vec)
@@ -187,9 +187,9 @@ vector<T> scalarMul(const double& k, const vector<T>& vec)
     return res;
 }
 
-vector<double> scalarAdd(const double& k, const vector<double>& vec)
+vector<float> scalarAdd(const float& k, const vector<float>& vec)
 {
-    vector<double> res;
+    vector<float> res;
     for (auto& elem : vec)
     {
         res.push_back(k + elem);
@@ -197,10 +197,10 @@ vector<double> scalarAdd(const double& k, const vector<double>& vec)
     return res;
 }
 
-std::pair<double, double> rotate(double x, double y, double theta)
+std::pair<float, float> rotate(float x, float y, float theta)
 {
-    const double x_ = x * std::cos(theta) - y * std::sin(theta);
-    const double y_ = x * std::sin(theta) + y * std::cos(theta);
+    const float x_ = x * std::cos(theta) - y * std::sin(theta);
+    const float y_ = x * std::sin(theta) + y * std::cos(theta);
     return { x_, y_ };
 }
